requirements.txt
fastapi
uvicorn
groq
pdf2image
pillow
pydantic
python-dotenv
pytest
streamlit
streamlit-confetti